feat(transcribe_streaming): add --separate-channel/-sc support
- Split _stream_one_channel() from transcribe_streaming() for multi-channel reuse
- save_result() gains channel param; adds _channel0/_channel1 suffix when set
- main() handles per-channel transcription with tempdir pattern
- Add tempfile import; update docstring and config log line

+++ funasr/models/ct_transformer/model.py
@@ -280,7 +280,7 @@ class CTTransformer(torch.nn.Module):
             mini_sentence = cache_sent + mini_sentence
             mini_sentence_id = np.concatenate((cache_sent_id, mini_sentence_id), axis=0)
             data = {
-                "text": torch.unsqueeze(torch.from_numpy(mini_sentence_id), 0),
+                "text": torch.unsqueeze(torch.from_numpy(mini_sentence_id).long(), 0),
                 "text_lengths": torch.from_numpy(np.array([len(mini_sentence_id)], dtype="int32")),
             }
             data = to_device(data, kwargs["device"])
